// frontend/src/app/models/event.model.ts
export class Event {
    event_id?: any;
    moderator_id?: any;
    event_name?: String ;
    event_artist?: String;
    event_duration?: Number;
    event_start?: String ;
    event_end?: String ;
    event_location?: String;
    event_category?: String;
    event_rating?: Number | any;
    event_description?: String;
    event_picture?: String;
    event_status?: String;
}